Clean up Slack cloud bot server boundaries
