+++ server/proliferate/server/cloud/slack/worker/__init__.py
@@ -1,0 +1,1 @@
+"""Worker entry points for the Cloud Slack bot."""
